+++ tests/assets/input_files/Coating Example.coated
