+++ tests/assets/input_files/Double Henyey-Greenstein.material
