+++ tests/assets/input_files/GENERIC_CAR 2023R2.3D Texture.1.OPT3DMapping
