+++ tests/assets/input_files/Gegenbauer.material
