+++ tests/assets/input_files/KB.spectrum
