+++ tests/assets/input_files/Non-Scattering.material
