+++ tests/assets/input_files/TL L.3D Texture.2.OPT3DMapping
