+++ tests/assets/input_files/User-Defined.material
